Refactor doctor model tests
